requirements.txt
absl-py==2.1.0
accelerate==0.21.0
aiofiles==23.2.1
altair==5.3.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
asttokens==2.4.1
attrs==23.2.0
bitsandbytes==0.43.1
bleach==6.1.0
blinker==1.7.0
blis==0.7.11
braceexpand==0.1.7
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
contexttimer==0.3.3
contourpy==1.2.1
cycler==0.12.1
cymem==2.0.8
decorator==5.1.1
decord==0.6.0
distlib==0.3.8
einops==0.6.1
einops-exts==0.0.4
exceptiongroup==1.2.1
executing==2.0.1
fairscale==0.4.4
fastapi==0.110.2
ffmpy==0.3.2
filelock==3.13.4
fonttools==4.51.0
fsspec==2024.3.1
ftfy==6.2.0
gitdb==4.0.11
GitPython==3.1.43
gradio==4.16.0
gradio_client==0.8.1
grpcio==1.62.2
gym==0.10.9
h11==0.14.0
h5py==3.12.1
httpcore==0.17.3
httpx==0.24.0
huggingface-hub==0.22.2
identify==2.5.36
idna==3.7
ifcfg==0.24
imageio==2.34.0
imageio-ffmpeg==0.4.9
importlib_metadata==7.1.0
importlib_resources==6.4.0
iopath==0.1.10
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kaggle==1.6.12
kiwisolver==1.4.5
langcodes==3.3.0
lazy_loader==0.4
llvmlite==0.42.0
lru-cache==0.2.3
Markdown==3.6
markdown-it-py==3.0.0
markdown2==2.4.13
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.2.1
nltk==3.8.1
nodeenv==1.8.0
numba==0.59.1
numpy==1.23.4
numpy-quaternion==2023.0.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
openai-clip==1.0.1
opencv-python==4.7.0.68
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
orjson==3.10.1
packaging==24.0
pandas==2.2.2
parso==0.8.4
peft==0.10.0
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.0
plotly==5.21.0
portalocker==2.8.2
pre-commit==3.7.0
preshed==3.0.9
prompt-toolkit==3.0.43
protobuf==5.28.2
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.0.0
pycocoevalcap==1.2
pycocotools==2.0.7
pydantic==2.7.0
pydantic_core==2.18.1
pydeck==0.8.1b0
pydub==0.25.1
pyglet==2.0.15
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-magic==0.4.27
python-multipart==0.0.9
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
regex==2024.4.16
requests==2.31.0
rich==13.7.1
rouge-score==0.1.2
rpds-py==0.18.0
ruff==0.4.1
safetensors==0.4.5
salesforce-lavis==1.0.2
scikit-image==0.22.0
scikit-learn==1.2.2
scipy==1.13.0
segment-anything==1.0
semantic-version==2.10.0
sentence-transformers==2.7.0
sentencepiece==0.1.99
shellingham==1.5.4
shortuuid==1.0.13
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
starlette==0.37.2
streamlit==1.33.0
svgwrite==1.4.3
sympy==1.12
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
text-unidecode==1.3
thinc==8.2.3
threadpoolctl==3.4.0
tifffile==2024.4.18
timm==0.6.13
tokenizers==0.15.2
toml==0.10.2
tomlkit==0.12.0
toolz==0.12.1
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
transformers==4.37.2
triton==3.0.0
typer==0.9.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
virtualenv==20.25.3
wasabi==1.1.2
watchdog==4.0.0
wavedrom==2.0.3.post3
wcwidth==0.2.13
weasel==0.3.4
webdataset==0.2.86
webencodings==0.5.1
websockets==11.0.3
Werkzeug==3.0.2
yacs==0.1.8
zipp==3.18.1
